// src/shared/modules/construction/sources/notion/scopes/schema.ts
import { z } from 'zod'

export const scopeOrAddonSchema = z.object({
  id: z.string(),
  name: z.string(),
  entryType: z.string().default('Scope'),
  unitOfPricing: z.string().default('unit'),
  coverImageUrl: z.string().nullable().default(null),
  relatedTrade: z.string(),
  relatedScopesOfWork: z.array(z.string()).default([]).optional(),
})

export type ScopeOrAddon = z.infer<typeof scopeOrAddonSchema>